requirements.txt
Django==1.8.3
Pillow==2.9.0
argparse==1.2.1
django-embed-video==1.0.0
django-polymorphic==0.7.1
django-storages==1.1.8
django-taggit==0.16.3
pytz==2015.4
requests==2.7.0
whitenoise==2.0.2
wsgiref==0.1.2
